md/raid5: Fix livelock when array is both resyncing and degraded.
commit 26ac107378c4742978216be1005b7291b799c7b2 upstream.

Commit 11ddc644965ee781d63fa2247cc5b2a9bfe41783:
  md: When RAID5 is dirty, force reconstruct-write instead of read-modify-write.

Causes an RCW cycle to be forced even when the array is degraded.
A degraded array cannot support RCW as that requires reading all data
blocks, and one may be missing.

Forcing an RCW when it is not possible causes a live-lock and the code
spins, repeatedly deciding to do something that cannot succeed.

So change the condition to only force RCW on non-degraded arrays.

Reported-by: Manibalan P <pmanibalan@amiindia.co.in>
Bisected-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Tested-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Fixes: 11ddc644965ee781d63fa2247cc5b2a9bfe41783
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>